SystemCode/src/backend/pyproject.toml
[tool.poetry]
name = "upskill-backend"
version = "0.1.0"
description = ""
authors = ["JackC90"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
Flask-Cors = "^3.0.10"
python-dotenv = "^1.0.0"
SQLAlchemy = "^2.0.5"
coverage = "^7.2.1"
Flask-Migrate = "^4.0.4"
psycopg2 = "^2.9.5"
alembic = "^1.10.2"
bcrypt = "^4.0.1"
PyJWT = "^2.6.0"
pyrsistent = "^0.19.3"
Flask-RESTful = "^0.3.9"
Flask-Seeder = "^1.2.0"
pandas = "^2.0.0"
sentence-transformers = "^2.2.2"
flask-pydantic = "^0.11.0"

[tool.poetry.dev-dependencies]
autopep8 = "2.0.2"
flake8 = "*"
ipython = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
